Add files via upload
Morse code is a communication method that uses a series of dots (short signals) and dashes (long signals) to represent letters, numbers, and symbols. It enables text-based messages to be transmitted over telegraph, radio, or light signals, widely used in the past for long-distance communication, especially in maritime and military contexts.